testhelper: introduce shares in test-info.yml
Change the way we list shares in test-info.yml
This will be used to add additional functionality to tests.

eg:
shares:
  # share name export1
  export1:
    # If present, it means the share is pre-mounted
    path: /mnt/share/export1-cephfs-vfs
    backend:
      # If present backend filesystem
      name: cephfs.vfs
    # If present, username to perform the tests with on this share
    users:
      test2: x
    # If present, the hostname to use to connect to the test server
    server: hostname1
  export2:

This commit still supports the old format of test-info.yml. The new
changes deprecates the old test-info.yml format which can be removed
once the sit-environment has been updated.

Signed-off-by: Sachin Prabhu <sp@spui.uk>